app/core/security.py
import jwt
from fastapi import HTTPException
from app.core.config import settings

def verify_jwt_token(token: str) -> dict:
    try:
        decoded_payload = jwt.decode(
            token,
            settings.JWT_ACCESS_SECRET,
            algorithms=[settings.JWT_ALGORITHM]
        )
        return decoded_payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token has expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")
